fix(storybook/addon-utils): stop wrapping stories in React elements — framework-agnostic withUtilStyles
`withUtilStyles` is registered for EVERY renderer (react, svelte, lit)
via `@canonical/storybook-config`, but #807 made it wrap each story in a
React element (`createElement("div", …, StoryFn())`). Non-React
renderers cannot mount a React element returned by a decorator: Svelte's
DecoratorHandler crashed every story with "decorator_Component is not a
function" (#839), and Lit failed the same way.

The decorator is now renderer-agnostic again — no react import at all.
It returns `StoryFn()` untouched and instead applies the computed
classes and inline layout styles to the story's real mount container,
synchronously at decorate time:

- canvas:   `#storybook-root` — the story's direct parent;
- autodocs: `#story--<id>[--primary]-inner`, the per-story element
  `renderStoryToElement` mounts into (also the direct parent).

Both targets exist while the renderer is composing the story, so this
preserves the two properties the React wrapper existed for:

- no first-paint flash: classes are in place before the story's first
  paint (the pre-#807 post-paint useEffect made stories flash their
  un-gridded layout, then reflow);
- subgrid binding: classes land on the story's DIRECT parent in both
  views, so a story root using `grid-template-*: subgrid` (Cards group)
  binds to the decorator grid on the canvas AND the docs page.

Containers that persist across stories (`#storybook-root`) are reset on
every application (all producible classes and managed style props are
removed first), so no story inherits a previous story's modifiers.

Verified live on all three renderers with the built addon:
- svelte/ds-global: canvas + autodocs render, classes on the right
  containers, no page errors (previously: every story crashed);
- react/ds-global: Cards subgrid story — `#storybook-root` gets
  `grid responsive comfortable app` with 12 computed columns and the
  story root's `grid-template-columns` resolves to `subgrid`; all
  autodocs story containers carry the grid classes;
- lit/ds-prototype: canvas renders with classes applied, no errors.

Fixes #839
